+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import os
 import numpy as np
 import pytest
 from wordviz.loading import EmbeddingLoader
@@ -8,15 +9,18 @@ def sample_embeddings():
     np.random.seed(42)
     return np.random.randn(50, 384)
 
-'''
-@pytest.fixture(scope='module')
-def vis():
-    loader = EmbeddingLoader()
+@pytest.fixture(scope='session')
+def loader_static():
+    loader_static = EmbeddingLoader()
     # let pytest find file
     file_path = os.path.join(os.path.dirname(__file__), 'embedding_100words.txt')
-    loader.load_from_file(file_path, 'word2vec')
-    return Visualizer(loader)
-'''
+    loader_static.load_from_file(file_path, 'word2vec')
+    return loader_static
+
+@pytest.fixture(scope='session')
+def vis_static():
+    return Visualizer(loader_static)
+
 
 @pytest.fixture(scope='session')
 def vis():